Converts most calls of map to forEach
map is used to functionally 'mutate' an array and enable chainable
calls to it. None of these calls did this, they only used the way
map iterates over an array and calls a function on each element as
it does. This is better handled by forEach since forEach does not
expect a return value and also iterates over an array.