fixes issue in examples where view explore in the viewer was no longer working
Fixes issue in examples where view explore in the viewer was no longer working due to adding the region's parameter. It adds checks to default region to the continental US if no region is present.  This should address any potential issues with links not working too. Since anything created before the addition regions had to the continental US