adds buttons for map interactions
Adds buttons to the bottom of the map and the left side for adding areas, exporting the map, and toggling the display of the map layer list. The commit does not include any logic to do any of the actions. Also found an error with the `react-dom` version and will require an `npm install`.