styled identify button and other items
Other items styled are charts and chart tooltips. The chart tooltip needs to be worked on. For instance, the tooltip should fall on the bar's max height (with an arrow). Also need to have the tooltip be persistent. We will add a link to explain how the metric was calculated.

For the last item, there are missing bars—all the undefined and NaN should be represented on the chart (label) 0 is fine, so it shows up. And the bars need to be in a specific order.