Fixed a few stlye options and adjustments for screen size
fixed my lint errors added a new xxl break-point for very large screens which adjusted the home page but we may need more.  for any screen widths over ```xxl: 2000```

Update for map height on smaller screens so we can fit map and map actions together. The map was 100% height on smaller screens and it was not clear that you could change map layers and see graphs etc.